file rename and tab reorg
renamed files and reorganized tab layout
