+++ Money Mentor/FrontEnd/Dashboards/home_dashboard.py
@@ -4,7 +4,7 @@ from tkinter import ttk
 import categories_tab
 import monthly_tab          
 import yearly_display_tab      
-import data_display_tab
+import monthly_display_tab
 
 from app_theme import apply_theme
 
@@ -45,8 +45,8 @@ def main():
     # --- Tabs ---
     categories_tab.create_tab(notebook)
     monthly_tab.create_tab(notebook)          
-    yearly_display_tab.create_tab(notebook)   
-    data_display_tab.create_tab(notebook)
+    monthly_display_tab.create_tab(notebook)
+    yearly_display_tab.create_tab(notebook)  
 
     root.mainloop()
